requirements.txt
openai
streamlit
python-dotenv
Pillow
requests
mysql-connector-python
flask
python-dotenv